add: arquivo powerup pra integrar os dois powerups

+++ Game_Explodable.java
@@ -7,7 +7,7 @@ public class Game_Explodable extends Game_Object{
         super(X, Y);
     }
 
-     protected void explode(int time){
+     protected void explode(long time){
         state = EXPLODING;
         explosionStart = currentTime;
         explosionEnd = currentTime + time;
